+++ js/data.js
@@ -1,0 +1,59 @@
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const IN_OUT_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const PHOTOS = [
+  'duonguyen-8LrGtIxxa4w.jpg',
+  'hbrandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const DESCRIPTIONS = [
+  'Замечательное место.',
+  'Хороший вид.',
+  'Рядом море.',
+  'Доступный транспорт.',
+];
+
+const MAX_PRICE = 100000;
+
+const MAX_ROOM = 100;
+
+const MAX_QUEST =200;
+
+const MAX_LAT = 90;
+
+const MAX_LNG = 180;
+
+const MIN_LAT_LOCATION = 35.65000;
+
+const MAX_LAT_LOCATION = 35.70000;
+
+const MIN_LNG_LOCATION = 139.70000;
+
+const MAX_LNG_LOCATION = 139.80000;
+
+const DIGIT_LOCATION = 5;
+
+const MAX_TITLE = 10;
+
+export {TYPE, FEATURES,IN_OUT_TIME,PHOTOS,DESCRIPTIONS,MAX_PRICE,MAX_ROOM,MAX_QUEST,MAX_LAT,MAX_LNG,MIN_LAT_LOCATION,MAX_LAT_LOCATION,MIN_LNG_LOCATION,MAX_LNG_LOCATION,DIGIT_LOCATION,MAX_TITLE};
